Split out app bundle from browser pages
